# Conflicts:
#	src/main/java/com/jd2/moviebase/controller/GenreController.java
#	src/main/java/com/jd2/moviebase/model/Genre.java
#	src/main/java/com/jd2/moviebase/service/GenreService.java
#	src/main/java/com/jd2/moviebase/util/ModelMapper.java

+++ src/main/java/com/jd2/moviebase/dto/GenreDto.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 @Builder
 public class GenreDto {
-    private int id;
-    private int tmdbId;
+    private Long id;
+    private Long tmdbId;
     private String name;
 
 }
